+++ src/eval_fns/ef07_sum_of_wells.rs
@@ -4,7 +4,7 @@ use crate::game::Board;
 pub struct SumOfWells;
 
 impl EvalFn for SumOfWells {
-    fn eval(&self, board: &Board) -> u8 {
+    fn eval(&self, board: &Board) -> u16 {
         (0..Board::WIDTH)
             .map(|col| calculate_well_depth(board, col))
             .sum()
